rename: change component name to camel case
