+++ components/article-card/ArticleCard.module.scss
@@ -1,15 +1,17 @@
-@use "styles/typography.scss" as typo;
+@use "../../styles/typography.scss" as typo;
 
 .wrapper {
   width: 100%;
   height: 100%;
 
   padding-bottom: 1rem;
+  border-radius: 10px;
 
   display: flex;
   flex-direction: column;
   gap: 8px;
 
+  background-color: var(--background-10);
   color: var(--text-color-10);
   text-decoration: none;
