+++ components/main-page/highlight-list/LatestPost.tsx
@@ -2,7 +2,7 @@
 
 import styles from "./highlight-list.module.scss";
 import { PostDataProps } from "@/type";
-import ArticleCard from "@/components/article-card/article-card";
+import ArticleCard from "@/components/article-card/ArticleCard";
 import NavButton from "@/components/button/NavButton";
 import useFloatingEffect from "@/hooks/useFloatingEffect";
