+++ components/projects/list/project-list.tsx
@@ -3,7 +3,7 @@
 import { useState, useEffect } from "react";
 
 import styles from "./project-list.module.scss";
-import ArticleCard from "@/components/article-card/article-card";
+import ArticleCard from "@/components/article-card/ArticleCard";
 import { sortByNewestArticle } from "@/utils/article";
 import { ProjectDataProps } from "@/type";
